# docs/b2b-core/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PharmAssist B2B Core API (/v1)",
    "version": "0.1.0",
    "description": "The B2B /v1 surface for integrators: Tier-1 Core, the `clinical_only` base tier below it, and the Tier-2 routes gated at `clinical` (a tier-gated route documents its 403). Authenticate with the `X-API-Key` header over TLS. Errors render in the stable `{\"error\": {code, message, request_id}}` envelope. See docs/b2b-core/API.md for the narrative (auth, tiers, locations without ΗΔΥΚΑ credentials, error codes, consent attestation, the 609/ΕΟΠΥΥ rule, rate limits, tri-state coverage, pagination, sandbox identifiers)."
  },
  "servers": [
    {
      "url": "https://api.<your-domain>",
      "description": "Live (FT-10, TBD)"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/v1/patients/{patient_key}": {
      "get": {
        "tags": [
          "b2b-v1",
          "b2b-v1"
        ],
        "summary": "Get Patient",
        "description": "Patient demographics by AMKA or EKAA.",
        "operationId": "get_patient_v1_patients__patient_key__get",
        "parameters": [
          {
            "name": "patient_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Key"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Patient"
                }
              }
            }
          },
          "409": {
            "description": "`retrieval_unavailable` — this location has no ΗΔΥΚΑ credentials, so patient/prescription retrieval is not available. Upstream-free routes (drugs, safety, conditions) keep working."
          },
          "403": {
            "description": "`tier_required` — this endpoint requires the `core` tier or higher; the customer's entitlement tier ranks below it."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/patients/{patient_key}/insurances": {
      "get": {
        "tags": [
          "b2b-v1",
          "b2b-v1"
        ],
        "summary": "Get Patient Insurances",
        "description": "ΕΟΠΥΥ/fund coverage entries, incl. the fund's `eopyy` flag (D-9 probe).\n\nNOTE: ΗΔΥΚΑ supplies no numeric patient-level co-pay % on this payload\n(D-7/D-9, verified — docs/b2b-core/insurances-probe.md). Patient co-pay\nEXEMPTIONS ride the patient profile (`participationExceptions` on\nGET /v1/patients/{key}); drug-level participation lives on /v1/drugs\n(participationPct).",
        "operationId": "get_patient_insurances_v1_patients__patient_key__insurances_get",
        "parameters": [
          {
            "name": "patient_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Key"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientInsurancePayload"
                  },
                  "title": "Response Get Patient Insurances V1 Patients  Patient Key  Insurances Get"
                }
              }
            }
          },
          "409": {
            "description": "`retrieval_unavailable` — this location has no ΗΔΥΚΑ credentials, so patient/prescription retrieval is not available. Upstream-free routes (drugs, safety, conditions) keep working."
          },
          "403": {
            "description": "`tier_required` — this endpoint requires the `core` tier or higher; the customer's entitlement tier ranks below it."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/patients/{patient_key}/intolerances": {
      "get": {
        "tags": [
          "b2b-v1",
          "b2b-v1"
        ],
        "summary": "Get Patient Intolerances",
        "description": "Recorded intolerances from ΗΔΥΚΑ (standalone — B2C only embeds these in\nthe profile). Requires ΕΟΠΥΥ category + caller-attested consent.",
        "operationId": "get_patient_intolerances_v1_patients__patient_key__intolerances_get",
        "parameters": [
          {
            "name": "patient_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Key"
            }
          },
          {
            "name": "patientConsent",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Patientconsent"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/V1Intolerance"
                  },
                  "title": "Response Get Patient Intolerances V1 Patients  Patient Key  Intolerances Get"
                }
              }
            }
          },
          "409": {
            "description": "`retrieval_unavailable` — this location has no ΗΔΥΚΑ credentials, so patient/prescription retrieval is not available. Upstream-free routes (drugs, safety, conditions) keep working."
          },
          "403": {
            "description": "`tier_required` — this endpoint requires the `core` tier or higher; the customer's entitlement tier ranks below it. Or `forbidden` — this location's ΗΔΥΚΑ account is not in the ΕΟΠΥΥ category (upstream 609)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/patients/{patient_key}/medicine-history": {
      "get": {
        "tags": [
          "b2b-v1",
          "b2b-v1"
        ],
        "summary": "Get Patient Medicine History",
        "description": "Executed-prescription history (cross-pharmacy, by AMKA/EKAA). Keeps the\nupstream pagination envelope incl. `blocked: true` on upstream 609.",
        "operationId": "get_patient_medicine_history_v1_patients__patient_key__medicine_history_get",
        "parameters": [
          {
            "name": "patient_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Key"
            }
          },
          {
            "name": "patientConsent",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Patientconsent"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Size"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RxHistoryPage"
                }
              }
            }
          },
          "409": {
            "description": "`retrieval_unavailable` — this location has no ΗΔΥΚΑ credentials, so patient/prescription retrieval is not available. Upstream-free routes (drugs, safety, conditions) keep working."
          },
          "403": {
            "description": "`tier_required` — this endpoint requires the `core` tier or higher; the customer's entitlement tier ranks below it. Or `forbidden` — this location's ΗΔΥΚΑ account is not in the ΕΟΠΥΥ category (upstream 609)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/patients/{patient_key}/conditions": {
      "get": {
        "tags": [
          "b2b-v1",
          "b2b-v1"
        ],
        "summary": "List Patient Conditions",
        "operationId": "list_patient_conditions_v1_patients__patient_key__conditions_get",
        "parameters": [
          {
            "name": "patient_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Key"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/V1Condition"
                  },
                  "title": "Response List Patient Conditions V1 Patients  Patient Key  Conditions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "b2b-v1",
          "b2b-v1"
        ],
        "summary": "Create Patient Condition",
        "description": "Record a condition for this patient AT THIS LOCATION. Conditions feed\nthe safety engine on POST /v1/safety/check. 409 when an active row for the\nsame (patient, condition) already exists at this location.",
        "operationId": "create_patient_condition_v1_patients__patient_key__conditions_post",
        "parameters": [
          {
            "name": "patient_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Key"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientConditionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Condition"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/patients/{patient_key}/conditions/{condition_id}": {
      "patch": {
        "tags": [
          "b2b-v1",
          "b2b-v1"
        ],
        "summary": "Update Patient Condition",
        "operationId": "update_patient_condition_v1_patients__patient_key__conditions__condition_id__patch",
        "parameters": [
          {
            "name": "patient_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Key"
            }
          },
          {
            "name": "condition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Condition Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientConditionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Condition"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "b2b-v1",
          "b2b-v1"
        ],
        "summary": "Delete Patient Condition",
        "description": "Soft delete — the row stays for audit, drops out of lists + safety checks.",
        "operationId": "delete_patient_condition_v1_patients__patient_key__conditions__condition_id__delete",
        "parameters": [
          {
            "name": "patient_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Key"
            }
          },
          {
            "name": "condition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Condition Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Condition"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientInsurancePayload": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "memberType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InsuranceMemberType"
              },
              {
                "type": "null"
              }
            ]
          },
          "ama": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ama"
          },
          "updateDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedate"
          },
          "directlyInsuredAmka": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directlyinsuredamka"
          },
          "fromEmaes": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fromemaes"
          },
          "lastActive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastactive"
          },
          "isRetired": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isretired"
          },
          "socialInsurance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InsuranceFund"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "PatientInsurancePayload"
      },
...
      "V1Intolerance": {
        "properties": {
          "activeSubstance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activesubstance"
          },
          "intolerance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intolerance"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "title": "V1Intolerance"
      },
...
      "V1Patient": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "amka": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amka"
          },
          "ekaa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ekaa"
          },
          "firstName": {
            "type": "string",
            "title": "Firstname"
          },
          "lastName": {
            "type": "string",
            "title": "Lastname"
          },
          "dateOfBirth": {
            "type": "string",
            "title": "Dateofbirth"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "sex": {
            "type": "string",
            "title": "Sex"
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "default": ""
          },
          "nationality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationality"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "participationExceptions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ParticipationException"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participationexceptions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "firstName",
          "lastName",
          "dateOfBirth",
          "age",
          "sex"
        ],
        "title": "V1Patient"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Per-location bearer key (`pa_live_…` / `pa_test_…`). TLS only. All failure modes return one indistinguishable 401."
      }
    }
  }
}